+++ packages/cubejs-backend-native/src/lib.rs
@@ -1,4 +1,5 @@
 #![feature(async_closure)]
+#![feature(thread_id_value)]
 
 extern crate findshlibs;