+++ packages/cubejs-backend-native/src/template/mod.rs
@@ -1,5 +1,8 @@
+#[cfg(feature = "python")]
+mod engine_python;
 mod entry;
 mod mj_value;
 mod neon;
+mod workers;
 
 pub use entry::template_register_module;