+++ rust/cubestore/cubestore/src/cachestore/cache_rocksstore.rs
@@ -715,6 +715,15 @@ impl CacheStore for RocksCacheStore {
         item: CacheItem,
         update_if_not_exists: bool,
     ) -> Result<bool, CubeError> {
+        if item.get_value().len() >= self.store.config.cachestore_cache_max_entry_size() {
+            return Err(CubeError::user(format!(
+                "Unable to SET cache with '{}' key, exceeds maximum allowed size for payload: {}, max allowed: {}",
+                item.key,
+                humansize::format_size(item.get_value().len(), humansize::DECIMAL),
+                humansize::format_size(self.store.config.cachestore_cache_max_entry_size(), humansize::DECIMAL),
+            )));
+        }
+
         self.cache_eviction_manager
             .before_insert(item.get_value().len() as u64)
             .await?;
@@ -1517,6 +1526,38 @@ mod tests {
         assert_eq!(row.expire.is_some(), true);
 
         RocksCacheStore::cleanup_test_cachestore("cache_set");
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_cache_set_max_entry_size() -> Result<(), CubeError> {
+        init_test_logger().await;
+
+        let config = Config::test("cachestore_set").update_config(|mut c| {
+            c.cachestore_cache_max_entry_size = 1 << 20;
+            c
+        });
+
+        let (_, cachestore) =
+            RocksCacheStore::prepare_test_cachestore("cache_set_max_entry_size", config);
+
+        let err = cachestore
+            .cache_set(
+                CacheItem::new(
+                    "prefix:key-with-wrong-size".to_string(),
+                    Some(60),
+                    "a".repeat(2 << 20),
+                ),
+                false,
+            )
+            .await;
+
+        assert_eq!(err, Err(CubeError::user(
+            "Unable to SET cache with 'key-with-wrong-size' key, exceeds maximum allowed size for payload: 2.10 MB, max allowed: 1.05 MB".to_string()
+        )));
+
+        RocksCacheStore::cleanup_test_cachestore("cache_set_max_entry_size");
 
         Ok(())
     }